Rename -> test_examples/test__examples.py, fix unit tests
